pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel>=0.37"]
build-backend = "setuptools.build_meta"

[project]
name = "Kafka_experiment"
version = "0.1.0"
description = "A Kafka producer-consumer implementation with PostgreSQL integration"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "MIT"}
authors = [
    {name = "Arun Warrier", email = "arunwarrier86@gmail.com"}
]
maintainers = [
    {name = "Arun Warrier", email = "arunwarrier86@gmail.com"}
]
keywords = ["kafka", "postgresql", "messaging", "streaming"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Python Modules",
]
dependencies = [
    "confluent-kafka==2.3.0",
    "psycopg2-binary==2.9.9",
    "python-dotenv==1.0.0",
    "six==1.16.0"
]

[project.scripts]
format = "black:main"  # Corrected entry point reference

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=22.0",
    "flake8>=6.1.0",
    "pytest-mock>=3.10.0",
]

[project.urls]
Homepage = "https://github.com/arunwarrier08/kafka_experiment"
Documentation = "https://github.com/arunwarrier08/kafka_experiment#readme"
Repository = "https://github.com/arunwarrier08/kafka_experiment.git"

[tool.setuptools]
packages = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

[tool.black]
line-length = 88
target-version = ["py312"]
